+++ virtcontainers/noop_network.go
@@ -5,17 +5,9 @@
 
 package virtcontainers
 
-import "context"
-
 // noopNetwork a.k.a. NO-OP Network is an empty network implementation, for
 // testing and mocking purposes.
 type noopNetwork struct {
-}
-
-// init initializes the network, setting a new network namespace for the Noop network.
-// It does nothing.
-func (n *noopNetwork) init(ctx context.Context, config NetworkConfig) (string, bool, error) {
-	return "", true, nil
 }
 
 // run runs a callback in the specified network namespace for
@@ -27,13 +19,13 @@ func (n *noopNetwork) run(networkNSPath string, cb func() error) error {
 
 // add adds all needed interfaces inside the network namespace the Noop network.
 // It does nothing.
-func (n *noopNetwork) add(sandbox *Sandbox, config NetworkConfig, netNsPath string, netNsCreated bool) (NetworkNamespace, error) {
-	return NetworkNamespace{}, nil
+func (n *noopNetwork) add(sandbox *Sandbox) error {
+	return nil
 }
 
 // remove unbridges and deletes TAP interfaces. It also removes virtual network
 // interfaces and deletes the network namespace for the Noop network.
 // It does nothing.
-func (n *noopNetwork) remove(sandbox *Sandbox, networkNS NetworkNamespace, netNsCreated bool) error {
+func (n *noopNetwork) remove(sandbox *Sandbox) error {
 	return nil
 }
